Replaced timeout override with timeout multiplier.
Instead of using testname_TIMEOUT=n to override, the user can use
testname_TIMEOUT_MULT=n to multiply the DEFAULT_TIMEOUT variable by
n for a timeout for a specific test.

This is better because overrides may break compilation on computer with
different speeds. Because of this, in a later patch, there will be no
in-file DEFAULT_TIMEOUT, but will be imported from the TIMEOUT shell
environment, so that each user sets a base timeout depending on his own
computer.