Changed defaults for timeout.
* Timeout is now set from the shell by exporting the TIMEOUT variable.
* There is no more in-file DEFAULT_TIMEOUT anymore.
* Timeout is now defaults to 0.

This way, users can opt-in for infinite loop breaking only if they
want by simply exporting their TIMEOUT environment variable with a value
in seconds.

It is better to have timeouts disabled because a suitable default value
is very difficult to nail down, and thus, it can break tests on lower-end
devices. Now each user opts-in and with the value they want.