Added Cppcheck support.
Cppcheck is turned on by default. User can turn it off
by doing one of the following:
* By setting SKIP_CPPCHECK=y in the environment and exporting it.
  USeful for temporarily disabling Cppcheck.
* By creating a RELEASE file in the source tree. Useful to disable
  Cppcheck permanently, for example, in a release tree.

Cppcheck unavailability will also be automatically detected and,
as a consequence, cppcheck will be disabled accordingly.

The CPPCHECK_EXTRA variable can be used to append options to
Cppcheck. This is useful for specifying suppressions.