Allow choosing an alternate compiler
Instead of hardcoding gcc as the compiler, use $(CC). Yes, we still
depend on that compiler supporting the options but maybe there are
cases where we *do* have other compilers that support GCC's options.
For example, clang is mostly compilers, but the most prominent case
is cross-compilation. In this situation the compiler is gcc but it
is usually called something other than 'gcc'.

Note that CC cannot be set from the environment. To override it
use 'make CC=...' instead.