Introduce the "strict" and "check" targets
Usually, by using Caddeus, targets are created in this order if no target
is specified:

1. Compile all objects
2. Compile application
3. Compile and run all unit tests
4. Run all scripted integration tests
5. Compile and run all compiled integration tests

This was never the original intention. The original intention was to test
every unit as soon as it is successfully compiled, like this:

1a. Compile object for module 1
1b. Compile and run unit test for module 1
2a. Compile object for module 2
2b. Compile and run unit test for module 2
... so on
3. Compile application
4. Compile and run all compiled integration tests
5. Run all scripted integration tests

Also observe how 4 and 5 are swapped, but that is a minor change and has
nearly no impact.

Additionally sometimes there is a need to *not* run any test at all, such
as when running an automatic rebuild for a distribution or when an end user
builds the program or when a developer just wants to check whether the
application fully builds. This was usually achieved by creating a RELEASE
file, but this is just an ugly hack. People expect "make" to just build the
application.

So, to be flexible and allow for the two ways of working, this change
introduces the "strict" target, which builds in the objectN+unittestN
order before building the final binary. This should be invoked like this:

make strict

This way, the default will be to just build the application. If a developer
wants to automatically run the test suite *after* the application has been
fully built, this can still be easily achieved either of the follwing:

make && make check
make check

(yes, "check" will depend on $(APP); it would not make sense otherwise)

The following changes were made:
* The default goal is now $(APP) instead of tested_$(APP).
* "strict" depends on $(OBJ_TESTS_TS) $(APP_TESTS_TS) $(APP_TESTS_TTS) $(APP)
* "check" depends on $(APP) $(OBJ_TESTS_TS) $(APP_TESTS_TS) $(APP_TESTS_TTS)
* $(APP) will no longer depend on any tests.

This seems to make much more sense.