Removed automatic object loading for unit tests
This turned out to be an incorrect approach. Unit tests should be able
to access all internal unit functions, including those not meant to be
exposed.

What happens right now (the wrong approach) is that Caddeus
automatically adds the module object (unit.o) into the test linking
process. This avoids having to reprocess the unit each time the test is
compiled but, because the module is added as an object and only public
prototypes are read, there is no way to access private functions.

This commits removes the automatic inclusion the unit object in the
linking process. The test should #include "unit.c" at the very top,
followed by #include "unit.h" right after. This way, there is full
access to internal functions, while prototype mismatches are detected.
This means, though, that if multiple test programs are created for a
single unit, the unit will be compiled along each test program, always.
This will make compilation slower and waste resources, but it seems to
be better than being unable to test the program.

I am open to suggestions on how to come up with a middle ground. I
dislike having to set variables for each test, though.